+++ src/functions.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 use Bakame\Stackwatch\AggregatorType;
 use Bakame\Stackwatch\Metrics;
-use Bakame\Stackwatch\Profiler;
 use Bakame\Stackwatch\Report;
 use Bakame\Stackwatch\Result;
+use Bakame\Stackwatch\Stack;
 
 if (!function_exists('stack')) {
     /**
@@ -16,22 +16,34 @@ if (!function_exists('stack')) {
      */
     function stack(callable $callback): Result
     {
-        return Profiler::execute($callback);
+        return Stack::execute($callback);
     }
 }
 
-if (!function_exists('stack_watch')) {
+if (!function_exists('stack_dump')) {
+    /**
+     * Profile a callable, dump the generated Span and return the callback returned value.
+     *
+     * The function returns the Result object generated for further usage if needed.
+     */
+    function stack_dump(callable $callback): Result
+    {
+        $result = Stack::execute($callback);
+        $result->span->dump();
+
+        return $result;
+    }
+}
+
+if (!function_exists('stack_dd')) {
     /**
      * Profile a callable, dump the generated Span and return the callback returned value.
      *
      * The function returns the callable expecged value
      */
-    function stack_watch(callable $callback): mixed
+    function stack_dd(callable $callback): never
     {
-        $result = Profiler::execute($callback);
-        $result->span->dump();
-
-        return $result->returnValue;
+        Stack::execute($callback)->span->dd();
     }
 }
 
@@ -46,11 +58,11 @@ if (!function_exists('stack_report')) {
      */
     function stack_report(callable $callback, int $iterations = 3, int $warmup = 0): Report
     {
-        return Profiler::report($callback, $iterations, $warmup);
+        return Stack::report($callback, $iterations, $warmup);
     }
 }
 
-if (!function_exists('stack_dump')) {
+if (!function_exists('stack_rdump')) {
     /**
      * Profile a callable, dump and return the generated @see Report.
      *
@@ -61,16 +73,13 @@ if (!function_exists('stack_dump')) {
      *
      * @throws Throwable
      */
-    function stack_dump(callable $callback, int $iterations = 3, int $warmup = 0): Report
+    function stack_rdump(callable $callback, int $iterations = 3, int $warmup = 0): Report
     {
-        /** @var Report $report */
-        $report = Profiler::dump($callback, $iterations, $warmup);
-
-        return $report;
+        return Stack::dumpReport($callback, $iterations, $warmup);
     }
 }
 
-if (!function_exists('stack_dd')) {
+if (!function_exists('stack_rdd')) {
     /**
      * Profile a callable, dump the generated @see Report and halt script execution.
      *
@@ -79,9 +88,9 @@ if (!function_exists('stack_dd')) {
      *
      * @throws Throwable
      */
-    function stack_dd(callable $callback, int $iterations = 3, int $warmup = 0): never
+    function stack_rdd(callable $callback, int $iterations = 3, int $warmup = 0): never
     {
-        Profiler::dd($callback, $iterations, $warmup);
+        Stack::ddReport($callback, $iterations, $warmup);
     }
 }
 
@@ -96,7 +105,7 @@ if (!function_exists('stack_metrics')) {
      */
     function stack_metrics(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): Metrics
     {
-        return Profiler::metrics($callback, $iterations, $warmup, $type);
+        return Stack::metrics($callback, $iterations, $warmup, $type);
     }
 }
 
@@ -113,10 +122,7 @@ if (!function_exists('stack_mdump')) {
      */
     function stack_mdump(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): Metrics
     {
-        /** @var Metrics $stats */
-        $stats = Profiler::dump($callback, $iterations, $warmup, $type);
-
-        return $stats;
+        return Stack::dumpMetrics($callback, $iterations, $warmup, $type);
     }
 }
 
@@ -131,6 +137,6 @@ if (!function_exists('stack_mdd')) {
      */
     function stack_mdd(callable $callback, int $iterations = 3, int $warmup = 0, AggregatorType $type = AggregatorType::Average): never
     {
-        Profiler::dd($callback, $iterations, $warmup, $type);
+        Stack::ddMetrics($callback, $iterations, $warmup, $type);
     }
 }